Created calculator model and moved mock there
